requirements.txt
Django==1.10.4
celery==3.1.20  
djangorestframework
psycopg2==2.5.3  
redis==2.10.5
numpy
PyDictionary
django-ajax-selects